NeuroTwin/requirements.txt
streamlit
plotly
pandas
numpy
pyttsx3
openpyxl
fpdf2
